+++ src/context/ThemeContext/ThemeContext.ts
@@ -1,0 +1,10 @@
+import { createContext } from "react";
+
+export interface ThemeContextValue {
+  theme: string;
+  toggleTheme: () => void;
+}
+
+export const ThemeContext = createContext<ThemeContextValue | undefined>(
+  undefined,
+);
