+++ src/context/ThemeContext/ThemeProvider.tsx
@@ -1,0 +1,20 @@
+import React, { useState, PropsWithChildren } from "react";
+import { useCallback } from "../../@lib";
+import { ThemeContext } from "./ThemeContext";
+
+export const ThemeProvider: React.FC<PropsWithChildren> = ({ children }) => {
+  const [theme, setTheme] = useState("light");
+
+  const toggleTheme = useCallback(() => {
+    setTheme((prevTheme) => (prevTheme === "light" ? "dark" : "light"));
+  }, []);
+
+  const value = {
+    theme,
+    toggleTheme,
+  };
+
+  return (
+    <ThemeContext.Provider value={value}>{children}</ThemeContext.Provider>
+  );
+};
