+++ src/context/ThemeContext/index.ts
@@ -1,0 +1,2 @@
+export * from "./ThemeProvider";
+export * from "./useThemeContext";
